Near end of first prototype
Because of being overwhelmed by regularly spawning enemies, I've replaced the prefab spawners with just their sprites and started a series of triggers that spawn enemies off-screen who then proceed towards the player. Also built in a "Game Over" screen that allows you to restart the level as well as a Title Screen that lets you start the game.