Title Screen Scroll
Added the scrolling background to the title screen and a script that has the camera constantly move right. Rather than using the tiles, I took the ground sprites directly from the source but it seems as though they are not being rearranged at the right Y-coordinates, so one of the ground sprites will appear in the right X-position, but they appear much higher.